+++ lib/src/model/scanner_exception.dart
@@ -1,0 +1,7 @@
+class ScannerException implements Exception {
+  final String _message;
+
+  String get message => _message;
+
+  ScannerException(this._message);
+}