+++ lib/src/utils/string_extension.dart
@@ -1,13 +1,3 @@
 extension StringExtension on String {
-  String clean() {
-    return replaceAll(' ', '')
-        .replaceAll('|', '')
-        .replaceAll('[', '')
-        .replaceAll(']', '')
-        .replaceAll('{', '')
-        .replaceAll('}', '')
-        .replaceAll('(', '')
-        .replaceAll(')', '')
-        .toLowerCase();
-  }
+  String clean() => replaceAll(RegExp(r'\D'), '');
 }