+++ lib/src/widget/camera_widget.dart
@@ -1,20 +1,20 @@
 import 'package:camera/camera.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:google_ml_kit/google_ml_kit.dart';
+import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 
 class CameraWidget extends StatefulWidget {
   final CameraController cameraController;
   final CameraDescription cameraDescription;
   final int scannerDelay;
 
-  const CameraWidget(
-      {Key? key,
-      required this.cameraController,
-      required this.cameraDescription,
-      required this.onImage,
-      required this.scannerDelay})
-      : super(key: key);
+  const CameraWidget({
+    Key? key,
+    required this.cameraController,
+    required this.cameraDescription,
+    required this.onImage,
+    required this.scannerDelay,
+  }) : super(key: key);
 
   final Function(InputImage inputImage) onImage;
 
@@ -25,11 +25,19 @@ class CameraWidget extends StatefulWidget {
 class CameraViewState extends State<CameraWidget> {
   int _lastFrameDecode = 0;
 
-  Future<void> stopCameraStream() async =>
-      widget.cameraController.stopImageStream();
+  Future<void> stopCameraStream() async {
+    if (!widget.cameraController.value.isStreamingImages) {
+      return;
+    }
+    return widget.cameraController.stopImageStream();
+  }
 
-  Future<void> startCameraStream() async =>
-      widget.cameraController.startImageStream(_processCameraImage);
+  Future<void> startCameraStream() async {
+    if (widget.cameraController.value.isStreamingImages) {
+      return;
+    }
+    return widget.cameraController.startImageStream(_processCameraImage);
+  }
 
   @override
   void initState() {
@@ -40,14 +48,13 @@ class CameraViewState extends State<CameraWidget> {
   @override
   Widget build(BuildContext context) {
     final mediaSize = MediaQuery.of(context).size;
-    final scale =
-        1 / (widget.cameraController.value.aspectRatio * mediaSize.aspectRatio);
-    return ClipRect(
-      clipper: _MediaSizeClipper(mediaSize),
-      child: Transform.scale(
-        scale: scale,
-        alignment: Alignment.topCenter,
-        child: CameraPreview(widget.cameraController),
+    return Transform.scale(
+      scale: 1 / mediaSize.aspectRatio,
+      child: Center(
+        child: AspectRatio(
+          aspectRatio: mediaSize.width / mediaSize.height,
+          child: CameraPreview(widget.cameraController),
+        ),
       ),
     );
   }
@@ -67,13 +74,13 @@ class CameraViewState extends State<CameraWidget> {
     final Size imageSize =
         Size(image.width.toDouble(), image.height.toDouble());
 
-    final imageRotation = InputImageRotationMethods.fromRawValue(
+    final imageRotation = InputImageRotationValue.fromRawValue(
             widget.cameraDescription.sensorOrientation) ??
-        InputImageRotation.Rotation_0deg;
+        InputImageRotation.rotation0deg;
 
     final inputImageFormat =
-        InputImageFormatMethods.fromRawValue(image.format.raw) ??
-            InputImageFormat.NV21;
+        InputImageFormatValue.fromRawValue(image.format.raw) ??
+            InputImageFormat.nv21;
 
     final planeData = image.planes.map(
       (Plane plane) {
@@ -96,21 +103,5 @@ class CameraViewState extends State<CameraWidget> {
         InputImage.fromBytes(bytes: bytes, inputImageData: inputImageData);
 
     widget.onImage(inputImage);
-  }
-}
-
-class _MediaSizeClipper extends CustomClipper<Rect> {
-  final Size mediaSize;
-
-  const _MediaSizeClipper(this.mediaSize);
-
-  @override
-  Rect getClip(Size size) {
-    return Rect.fromLTWH(0, 0, mediaSize.width, mediaSize.height);
-  }
-
-  @override
-  bool shouldReclip(CustomClipper<Rect> oldClipper) {
-    return true;
   }
 }