+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.2
+
+* Update readme.md.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Update readme.md.
+* Update dependencies.
+* Add functions to pause or resume controller and listener.
+* Disable/enable camera stream resume/pause.
+* Lock orientation.
+
+